Coding Days: level type changed to int
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2761 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/finance/application/BlackScholesHullWhiteSolver.hpp
@@ -106,7 +106,7 @@ public:
 	 */
 	virtual ~BlackScholesHullWhiteSolver();
 
-	void constructGrid(sg::base::BoundingBox& myBoundingBox, size_t level);
+	void constructGrid(sg::base::BoundingBox& myBoundingBox, int level);
 
 	/**
 	 * In order to combine the Black Scholes Equation with the Hull White Equation you have to provided
