+++ src/sgpp/pde/application/PoissonEquationSolver.cpp
@@ -37,7 +37,7 @@ PoissonEquationSolver::~PoissonEquationSolver()
 	}
 }
 
-void PoissonEquationSolver::constructGrid(BoundingBox& BoundingBox, size_t level)
+void PoissonEquationSolver::constructGrid(BoundingBox& BoundingBox, int level)
 {
 	this->dim = BoundingBox.getDimensions();
 	this->levels = level;
